add view and edit of transfusion center and admin profile

+++ ISA - project/src/main/java/com/javaguide/ISAprojekat/model/User.java
@@ -17,7 +17,7 @@ public class User implements UserDetails {
 @Id
 @SequenceGenerator(name = "my_seq_gen_user", sequenceName = "my_seq_gen_user", initialValue = 1, allocationSize = 1)
 @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "my_seq_gen_user")
-    private Long id;
+    private Integer id;
 
     @Column
     private String password;
@@ -97,7 +97,7 @@ public class User implements UserDetails {
         return false;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -142,7 +142,7 @@ public class User implements UserDetails {
         this.role = role;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
@@ -182,4 +182,5 @@ public class User implements UserDetails {
     public Role getRole() {
         return role;
     }
+
 }
